The OLAP objects are now SPObjects. The OLAPObject that all of the objects in MondrianModel extended was very close to SPObject already. Most of these changes were small updates to method names or listeners to make them match the SPObject ones.
The undo events was left in for the OLAPObjects. In the future we will want an undo pattern within the SPObjects.

The getDependencies and removeDependencies for OLAPObjects currently throw exceptions because they are referencing dependencies by name not by object type. This makes finding dependencies more difficult. This will be addressed shortly.