NEW - bug 2835: Dragging a PK column to the FK table causes an IOOBE http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2835
Fixed an old off-by-one error. If there are three children (indexed at 0, 1, and 2) and the insertion point is at position 2
we are looking to insert an object after child at index 1. Removing child at index 2 should not affect the insertion point
as it comes just after the point we are looking at.