NEW - bug 2458: Create Critic Manager http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2458
Added in CriticBadges to the play pen. This allows objects that have criticisms on them to appear in the play pen with a badge beside them.
The idea of a transient child has been added to the PlayPenConetentPane. At current it is a type of SPObject that has a correct parent pointer but changes do not fire events, which would lead to it being persisted, or gets returned in getChildren. We should see if there are other places where these kinds of child types may be of interest and formalize how transient children work.
A method for getting a UI point for model objects has been added to the UI components. This gives each component that has an external badge the chance to say where the badge should go. For relationships this ends up in the middle of the line which can cover some of a relationship's label which could be improved.