Added to the DBTreeModel so that it can be made optional to display the playpen database, columns, relationships, imported keys, and indices. This new functionality of the DBTreeModel is used in the newly added source column chooser in ColumnEditPanel.
Refactored the SQLTypeTreePopupAction into the library as PopupJTreeAction so that the ColumnEditPanel can use it for both the source column chooser and the data type chooser.

The data type chooser now updates the data type button text properly if a non-UserDefinedSQLType is selected in the JTree.

Fixed several NPE's with the column's auto increment being null.

Also, fixed some Eclipse warnings in the DBTree class.